Minor changes to UI

+++ food_ordering_app/lib/features/admin/order/domain/entities/orderitem.dart
@@ -6,7 +6,7 @@ class OrderItem extends Equatable {
   String restaurantId;
   DateTime orderDate;
   int totalAmount;
-  int? ratingGiven;
+  double? ratingGiven;
   String status;
   List<Map<String, dynamic>> order;
